Refactor UpdateHandlerScope
